Increase minReplyDelay of DDC request
Tested with DELL U3818DW connected via USB-C to a MacBook Pro 15 (2017)
with a Radeon Pro 560 and Intel HD Graphics 630 graphics card.

The value of 30 * kMillisecondScale was reached by experimenting, 20
fails ~50% of the time whereas 30 succeeds 100% of the time (on the
first try).

This magic number seems to have been discovered previously, albeit in
relation to ATI/ADM, at:
https://www.tonymacx86.com/threads/controlling-your-monitor-with-osx-ddc-panel.90077/page-3#post-581242